requirements.txt
fastapi
uvicorn
httpx
python-dotenv
jsonpath-ng
